Replication dataset from Horowitz and Klaus (2020)

Horowitz, J., Klaus, K. "Can Politicians Exploit Ethnic Grievances? An Experimental Study of Land Appeals in Kenya." Political Behavior 42, 35–58 (2020). https://doi.org/10.1007/s11109-018-9485-1

Usage

land_experiment

Format

A data frame with 375 rows and 9 variables:

support

Support for the hypothetical candidate making the appeal

land_insecure

Moderator of interest measuring if the respondent is land insecure

treat_comb

Treatment indicator for land-based appeals

prepost

Indicator for if the moderator was measured before (0) or after(1) treatment

age

Age of the respondent

female

Indicator for if the respondent identifies as female

close_own

Indicator for if the respondent feels close to their own ethnic group

educ

Level of education (1=no schooling, 2 = some primary, 3 = complete primary, 4 = some secondary, 5 = completed secondary, 6 = college, 7 = some university, 8 = completed university)

,
treat

Original treatment variable indicating control (0),land-based appeal (1), and land- and ethnic-based appeals (2).

Source

doi:10.7910/DVN/XBWR8N

Details

A dataset containing the results of an experimental study of political appeals about land greivances on candidate support in Kenya