Run pre-treatment bounds.

Run pre-treatment bounds.

Usage

pre_bounds(formula, data, moderator, conf_level = 0.95, outcome_mono = 1L)

Arguments

formula

A formula with syntax y ~ t, where y is the (unquoted) name of the outcome and t is the (unquoted) name of the treatment.

data

A data.frame containing variables in the formula, moderator, and covariates arguments.

moderator

A one-sided formuala with syntax ~ d, where d is the (unquoted) name of the moderator variable for the CATE.

conf_level

A numeric indicating the confidence level for the bootstrap confidence intervals.

outcome_mono

A integer indicating the direction of the priming monotonicity assumption. The default value 1 indicates that asking the moderator question in the pre-test moves outcomes in a positive direction for all units. The value -1 indicates it moves outcomes in a negative direction for all units.

Value

A list object containing bounds.

Examples

data(delponte)
pre_bounds(
  formula = angry_bin ~ t_commonality,
   data = delponte,
  moderator = ~ itaid_bin
)
#> $lower
#>            
#> -0.5923203 
#> 
#> $upper
#>           
#> 0.3221525 
#> 
#> $ci_lower
#> [1] -0.6875343
#> 
#> $ci_upper
#> [1] 0.4035053
#> 
#> $pre_est
#> [1] -0.2701678
#>