Run Gibbs sampler for the random moderator placement design

Run Gibbs sampler for the random moderator placement design

Usage

prepost_gibbs(
  formula,
  data,
  prepost,
  moderator,
  covariates,
  iter = 1000,
  thin = 1,
  burn = 0,
  offset = 0,
  monotonicity = TRUE,
  stable = TRUE,
  saturated = TRUE,
  priors
)

Arguments

formula

A formula with syntax y ~ t, where y is the (unquoted) name of the outcome and t is the (unquoted) name of the treatment.

data

A data.frame containing variables in the formula, moderator, and covariates arguments.

prepost

A one-sided formula with syntax ~ z, where z is the indicator variable for whether the moderator was measured pre- or post-treatment.

moderator

A one-sided formuala with syntax ~ d, where d is the (unquoted) name of the moderator variable for the CATE.

covariates

A one-sided formula with syntax ~ x1 + x2, where the right-hand side variables signify which covariates the Gibbs will use to try and narrow the bounds.

iter

Integer indicating the number of iterations for the Gibbs sampler.

thin

Integer indicating how often the Gibbs sampler should save a draw.

burn

Integer indicating how many iterations should be performed before saving draws in the Gibbs sampler.

offset

A numeric value indicating the center of the prior distribution for the covariate coefficients.

monotonicity

A logical signifying whether the model assumes monotonicity.

stable

A logical signifying whether the model assumes that the the pre vs post indicator does not affect the moderator under the control condition for treatment.

saturated

A logical indicating whether the coefficients on the covariates are allowed to vary by the principal strata.

priors

A list object containing the priors for the Gibbs sampler. Priors include beta.precision, psi.precision, alpha, y.alpha, and y.beta.

Value

A list object containing Gibbs posterior quantities of interest and parameters.

Examples

data(land_experiment)
prepost_gibbs(
  support ~ treat_comb,
  data = land_experiment,
  moderator = ~ land_insecure,
  prepost = ~ prepost,
  covariates = ~ educ + age,
  iter = 50
)
#> $betas
#>       (Intercept)       educ          age   treat_comb     prepost
#>  [1,]    5.360947  0.7496439  0.199111833 -0.880491615 -3.04026633
#>  [2,]    6.714527 -0.8714747  0.015179307  1.895980741  1.17678890
#>  [3,]    3.603818 -1.2197017 -0.037954492 -0.727038254 -1.49949742
#>  [4,]    4.329912 -1.0216453 -0.043535796 -1.383183150 -2.12506999
#>  [5,]    4.661942 -0.7874893 -0.018695431 -2.703682322 -1.59698100
#>  [6,]    2.888665 -0.7753985  0.002136428 -1.139469597  0.23131052
#>  [7,]    3.456364 -0.6561641  0.001429577 -0.154246841 -1.04199287
#>  [8,]    3.439634 -0.4932303 -0.002740113  0.233083104 -1.09324380
#>  [9,]    3.852158 -0.5729627 -0.013327202  1.014437446 -0.30406544
#> [10,]    2.608060 -0.3221372 -0.001654098  0.460988550  0.18890387
#> [11,]    3.531512 -0.4102868 -0.014713613  0.353572355 -0.08642819
#> [12,]    4.183591 -0.3549070 -0.036776005  0.524247351  0.38503920
#> [13,]    3.906570 -0.4870842 -0.021512545  0.953922691  0.18520519
#> [14,]    4.486802 -0.4918302 -0.015890503  0.228722024 -0.36309049
#> [15,]    3.918320 -0.3446895 -0.019037139  0.098968121 -0.05121159
#> [16,]    2.237825 -0.1846009 -0.016602714  1.007077130  1.08243165
#> [17,]    2.992241 -0.1760367 -0.018924983  0.179090823  0.46482387
#> [18,]    3.886705 -0.3650398 -0.017437826  0.489927203  0.03373340
#> [19,]    4.195434 -0.4776127 -0.005108270  0.709705666 -0.42909845
#> [20,]    3.807008 -0.4658234 -0.004802488  1.053860439 -0.29129125
#> [21,]    3.857503 -0.2863276 -0.014720723  0.498700942  0.10533767
#> [22,]    4.948322 -0.5115256 -0.018862937  0.377871138 -0.05084449
#> [23,]    4.310637 -0.4744525 -0.018021618  0.716041936  0.57531299
#> [24,]    4.866601 -0.3869392 -0.035247121  0.033869985 -0.02119845
#> [25,]    5.229875 -0.4509056 -0.037299195 -0.194935150 -0.03658707
#> [26,]    4.991199 -0.5222834 -0.034247806  0.646719002  0.69101428
#> [27,]    2.967274 -0.4404479 -0.008563264  1.453252193  1.19846629
#> [28,]    2.445298 -0.2740878 -0.002319318  0.997870513  0.27379740
#> [29,]    3.260362 -0.3928394 -0.003472819  0.602114066  0.18101252
#> [30,]    3.912892 -0.3519501 -0.017222998  0.163619125 -0.44926131
#> [31,]    4.153306 -0.3229375 -0.023142783 -0.343951559  0.17714056
#> [32,]    4.180858 -0.4255618 -0.026774770  0.342679002  0.10997527
#> [33,]    3.564354 -0.3502674 -0.020395209  0.265361940  0.08314498
#> [34,]    4.376807 -0.4180846 -0.030489212  0.781994739 -0.16815800
#> [35,]    3.722347 -0.2772692 -0.006631316 -0.382254406 -0.75787283
#> [36,]    4.076711 -0.3841140 -0.017984278 -0.357254609 -0.47179245
#> [37,]    3.886690 -0.2696448 -0.009747825 -0.872833059 -1.01196561
#> [38,]    4.258991 -0.3246719 -0.031963127 -0.108704469 -0.27600809
#> [39,]    4.479448 -0.4191827 -0.028199230  0.076520091 -0.03522273
#> [40,]    4.582739 -0.3058748 -0.018975827 -0.514655439 -0.72725236
#> [41,]    3.923768 -0.2908268 -0.018348379  0.008104617 -0.59757639
#> [42,]    4.327726 -0.2845248 -0.018465207 -0.676603906 -0.95411346
#> [43,]    3.166889 -0.2064517  0.009343143 -0.117050488 -0.91000514
#> [44,]    4.105548 -0.2702820 -0.019812702 -0.274852030 -0.97558549
#> [45,]    4.355335 -0.3636126 -0.015863134 -0.154373866 -0.47663151
#> [46,]    4.809209 -0.4697346 -0.025356329  0.256147203  0.11176420
#> [47,]    4.847277 -0.4140777 -0.027827286 -0.164341310 -0.63880817
#> [48,]    5.267361 -0.4335057 -0.031873993 -0.170242763 -0.45402553
#> [49,]    5.198040 -0.4598486 -0.034333084 -0.359589837 -0.51636658
#> [50,]    3.865670 -0.2663263 -0.012473444 -0.240341564 -0.67307356
#>       treat_comb:prepost       s111       s100 treat_comb:s111 treat_comb:s100
#>  [1,]        7.167873237 -8.4724579  7.3099974      4.71032137       4.2063118
#>  [2,]        4.029491973 -3.9021044 12.0297851      2.57401695       1.4954315
#>  [3,]       10.668701448  2.5581941 19.1999546      2.36590676       2.2534755
#>  [4,]        8.856810558  0.8889463 16.1762592      2.16987305       3.6970858
#>  [5,]        6.313822895 -1.8729291 15.8371347      4.40775230       4.2172097
#>  [6,]        3.230513517 -0.3689113 14.9849090      1.89791018       5.3908529
#>  [7,]        1.693774274 -2.3657336 11.4715557      2.75487378       4.6593446
#>  [8,]        1.261210527 -0.8716101  8.5882015      0.33745337       5.3235180
#>  [9,]       -0.392926248 -1.4075869  9.9349022      0.83235352       3.1407092
#> [10,]       -0.010414596 -3.7022213  8.4823429      4.14186042       3.9326271
#> [11,]        0.362174191 -0.8563141  7.6428295     -1.04613953       3.6469506
#> [12,]       -0.823580153 -2.4689234  6.8688173      1.34565931       2.9706067
#> [13,]       -0.730495476 -0.4300353  8.4129075     -0.28471535       0.9595057
#> [14,]       -0.780487263 -1.7290088  6.0641847      1.38792676       2.4145619
#> [15,]       -0.590861558 -2.2660108  3.5453154      1.89839087       4.3240185
#> [16,]       -1.559294780 -1.2691653  6.0964465      0.81677580       3.0968318
#> [17,]       -0.922941541 -2.4874703  6.4088780      1.86248867       1.9783284
#> [18,]       -0.876519357 -1.8312803  7.1877352      0.17392555      -0.6739680
#> [19,]       -0.717099204 -1.9092122  5.7557637      0.20428391      -1.5794005
#> [20,]       -1.109849556 -1.2161418  3.8172080     -0.62949693       1.0847547
#> [21,]       -1.507705888 -1.4116722  6.8174744     -0.64881415      -2.2324662
#> [22,]       -0.672980637 -2.1845786  1.8721733     -0.02676394       2.4531559
#> [23,]       -1.906720183 -1.7154823  3.0315244     -0.87681170      -0.9360357
#> [24,]       -1.028372285 -4.1460860  2.5476552      4.37575398       2.0760578
#> [25,]       -0.400761114 -3.3573554  1.4318431      2.09933751       3.8364817
#> [26,]       -1.083258674 -2.7706163  2.5887199      1.25893498       1.3227017
#> [27,]       -1.590882362 -1.1730419  2.9254009     -1.67893505       1.7195773
#> [28,]       -1.068659750 -0.3206608  1.9737958     -1.49993018       3.0907766
#> [29,]       -0.796110069 -1.9729831  1.9586502      0.71916621       1.9122387
#> [30,]       -0.470997536 -1.0141708  4.0477372     -0.06718887       1.4342257
#> [31,]       -0.578170457 -2.6801781  3.0130879      0.98184860       2.8297680
#> [32,]       -0.424862775 -1.4744869  4.5253842      1.33988513       2.3099822
#> [33,]        0.012410287 -0.3970104  4.1569299     -0.36743325       2.3517212
#> [34,]       -0.823415423 -1.5409556  3.2307775     -0.14538325       0.7907857
#> [35,]        0.176022078 -1.4126312  3.3593832      1.16402650      -0.1018005
#> [36,]       -0.009652835 -2.3437229  1.3694301      2.61645770       1.6019303
#> [37,]        0.631833541 -2.3054986  0.2426254      2.42936854       2.9611904
#> [38,]       -0.492181515 -1.3090243  0.2649634      0.63690252       2.7011017
#> [39,]       -0.362760156 -2.7952177  0.8298639      1.54788155       3.0841085
#> [40,]       -0.029289480 -3.1250368  1.4917796      1.61472830       1.2254280
#> [41,]       -0.450159171 -2.5703002  0.1970804      2.41334401       3.5115585
#> [42,]        0.092390935 -1.5316279 -1.9023171      0.34556802       5.9496794
#> [43,]       -0.605576459 -2.8749448 -0.4777297      0.84327700       2.0060643
#> [44,]        0.397987836 -2.4234397 -2.2792875      1.49169913       3.7091463
#> [45,]       -0.305131115 -0.6449853 -3.7542911     -0.54116584       4.7656308
#> [46,]       -0.982566733 -0.4885519 -3.0534556      0.31189270       3.0232693
#> [47,]       -0.204880922 -1.3471927 -2.2183581      2.62954265       2.6352843
#> [48,]       -0.470675859 -4.3986593 -1.5467762      5.00691300       3.1853404
#> [49,]       -0.020463119 -2.9418250 -1.9871704      1.63062768       3.4413358
#> [50,]       -0.246128494 -1.7638947 -1.9615075     -0.50080409       0.7416147
#>       prepost:s111 prepost:s100 treat_comb:prepost:s111 treat_comb:prepost:s100
#>  [1,]  -4.54077343 -0.563594400             10.90837981              0.94059415
#>  [2,]   1.09582061 -5.235717264              1.59381386              2.68404394
#>  [3,]   2.76045840 -6.136286438              1.94354248             -0.98263472
#>  [4,]   1.13586540 -3.497740189              1.29746961             -0.90128627
#>  [5,]   3.28212656 -6.614979174             -4.49654208              1.18617108
#>  [6,]   1.17627462 -5.999071937              4.13109611              2.17730199
#>  [7,]   0.38179395 -2.584663456              0.45815288             -1.24587243
#>  [8,]   1.90886505 -0.389360460             -2.08720490              0.29404154
#>  [9,]  -1.09277618 -3.602593375             -1.66190071             -0.01616376
#> [10,]   2.12269859 -2.225772783             -4.22661117             -1.18661525
#> [11,]   2.10968970  1.713467287              1.51210310             -3.14500094
#> [12,]   1.72524130  1.803831882             -0.64289765             -1.29253809
#> [13,]   1.26534865 -2.222498633             -2.48563442              0.80503581
#> [14,]   1.11536985  0.507079570             -2.74407497             -2.71035746
#> [15,]   1.58218017  1.783326671              0.03537488             -5.27891244
#> [16,]  -0.29282561 -2.424438760              5.43894976             -2.32047319
#> [17,]  -0.13527509 -2.677134835              4.23855683             -1.84588453
#> [18,]   0.72811587 -5.761531845             -4.81320792              3.07118648
#> [19,]   0.92210692 -1.814272112              2.02912123              0.51565262
#> [20,]   0.65573768 -0.006698406             -0.82339095             -1.58114497
#> [21,]   0.29181293 -4.553275055              3.59446270              3.22822282
#> [22,]  -0.36897274 -0.481118395             -0.11743942             -2.21911850
#> [23,]   1.97120346 -1.878994309              1.78490185              2.25804107
#> [24,]   2.55221825  0.541880983              0.31569062             -4.29678890
#> [25,]   1.42151206 -0.038732955              1.72433265             -4.93671940
#> [26,]   0.75804249 -1.389223712              4.79253783             -1.87215846
#> [27,]  -0.80747143 -1.209124323              5.37985714             -2.37819910
#> [28,]  -1.48790938 -0.758462546              5.64552412             -3.96541195
#> [29,]   0.64845471 -1.668098809              3.02193609              0.04952381
#> [30,]   1.55667584 -3.877053912              3.25113841             -0.40959352
#> [31,]   1.56953904 -1.228707555             -1.55329280             -2.17289906
#> [32,]   0.83901862 -4.973859392             -2.95428194             -1.59602143
#> [33,]   0.45729870 -4.448164551              1.85197270             -2.50280452
#> [34,]   1.26075753 -0.563142816             -0.11157724             -3.78080925
#> [35,]  -0.12957266  0.192629836              0.01428784             -2.10945754
#> [36,]  -0.40354418  1.620395266             -2.96511206             -2.41202361
#> [37,]   1.25657557  1.840117638             -2.81011007             -2.92981993
#> [38,]  -0.01411209  0.997123965              1.11116463             -0.64134832
#> [39,]   1.01835690  5.418637336              0.04365759             -6.43133424
#> [40,]   2.59906610 -1.109268329             -0.64973127              0.97619253
#> [41,]   0.55687070 -1.521341548             -0.43789462              1.81990202
#> [42,]   1.06674736  1.119264843              0.99525121             -0.57238991
#> [43,]   2.60766639 -0.632710839              1.99614151              3.80086499
#> [44,]   1.30853688  1.053793599             -0.29989813             -0.37136612
#> [45,]   0.93629045  0.401264268              0.58702787              2.38866394
#> [46,]  -0.84881895  0.909794139              4.94044310              2.99613180
#> [47,]   1.00372283  5.707885174              4.25242117             -2.62099627
#> [48,]   2.46967016  4.213904820              0.06982383             -2.81513942
#> [49,]   1.52279855  4.346143980              1.64401889             -2.89110794
#> [50,]  -0.41594699  3.298315709              2.43127897              0.18780024
#> 
#> $psis
#> $psis$s111
#>       (Intercept)          educ          age
#>  [1,]   -4.930791 -1.775707e+01 -1.709647575
#>  [2,]  -16.937355 -1.463235e+01 -1.187990176
#>  [3,]  -20.605408 -1.290646e+01 -0.839898196
#>  [4,]  -16.641018 -1.223794e+01 -0.729981238
#>  [5,]  -17.573579 -1.128451e+01 -0.502147725
#>  [6,]  -19.566128 -9.805020e+00 -0.347622719
#>  [7,]  -16.740244 -9.111698e+00 -0.232280550
#>  [8,]  -21.567732 -7.368487e+00 -0.103687328
#>  [9,]  -24.487560 -5.944618e+00  0.009977187
#> [10,]  -19.599389 -5.692360e+00 -0.033676939
#> [11,]  -17.917162 -5.156318e+00 -0.022704015
#> [12,]  -18.480911 -4.407352e+00  0.028379108
#> [13,]  -16.271399 -3.867590e+00  0.046367207
#> [14,]  -12.097064 -4.134123e+00  0.042253258
#> [15,]  -15.348618 -2.969216e+00  0.084148925
#> [16,]  -15.344483 -2.316298e+00  0.089851474
#> [17,]  -14.932350 -1.954656e+00  0.116813957
#> [18,]  -14.554829 -1.662970e+00  0.151799278
#> [19,]  -16.385666 -1.061163e+00  0.173240388
#> [20,]  -19.959739 -2.252623e-01  0.230500610
#> [21,]  -17.152863 -2.679732e-01  0.197239675
#> [22,]  -13.559912 -4.352659e-01  0.158150852
#> [23,]  -12.391973 -5.317492e-01  0.154562184
#> [24,]  -11.567622 -4.847561e-01  0.147905690
#> [25,]   -9.758542 -7.576124e-01  0.158385371
#> [26,]   -8.470929 -6.777807e-01  0.138451902
#> [27,]   -8.115154 -2.871098e-01  0.108339321
#> [28,]   -7.328825 -3.066888e-01  0.106439057
#> [29,]   -6.019044 -1.641870e-01  0.069736447
#> [30,]   -5.049540 -1.989936e-01  0.057582078
#> [31,]   -5.743910  4.732739e-02  0.057379012
#> [32,]   -4.997933 -4.659854e-02  0.049728632
#> [33,]   -4.393671 -4.371625e-02  0.039913397
#> [34,]   -5.365844  5.033166e-02  0.051925231
#> [35,]   -5.355648 -6.184357e-02  0.066727192
#> [36,]   -5.150964 -2.104963e-02  0.050871609
#> [37,]   -4.113169  5.117100e-03  0.024399033
#> [38,]   -4.813761  4.085078e-02  0.049487018
#> [39,]   -3.134781 -1.580112e-01  0.032401903
#> [40,]   -2.650138 -1.881319e-01  0.025114682
#> [41,]   -3.053820 -7.908971e-02  0.028487758
#> [42,]   -2.589771 -9.155068e-02  0.012807073
#> [43,]   -3.068741 -5.191935e-02  0.017071395
#> [44,]   -2.905475 -6.193064e-03  0.014054048
#> [45,]   -3.211287  8.334605e-04  0.015787931
#> [46,]   -3.520862  1.356842e-01  0.020771510
#> [47,]   -3.985732  1.511815e-01  0.030019961
#> [48,]   -3.176961  6.309308e-02  0.009135593
#> [49,]   -3.540515  1.487967e-01  0.002990984
#> [50,]   -4.492773  1.974119e-01  0.034227492
#> 
#> $psis$s100
#>       (Intercept)         educ          age
#>  [1,]  3.50931677  0.334565784  0.761926059
#>  [2,]  2.28531281  0.515095226  0.161010976
#>  [3,]  2.53139460 -0.063519048 -0.012791798
#>  [4,]  1.28832459 -0.136864539 -0.011932868
#>  [5,]  0.40083153 -0.022005800 -0.005580178
#>  [6,]  0.64700870 -0.032312268 -0.022562714
#>  [7,] -0.28648066  0.153470580 -0.017240829
#>  [8,] -0.02827326  0.100068560 -0.029486047
#>  [9,] -0.86617700  0.082734685 -0.014042959
#> [10,] -1.68585845  0.159867744 -0.008476409
#> [11,] -0.62359593 -0.070238113 -0.013688965
#> [12,] -1.47538251 -0.077366820  0.004998075
#> [13,] -1.00382127 -0.028537641 -0.011698018
#> [14,] -0.35510050 -0.082526509 -0.027826072
#> [15,] -0.21678030 -0.128277826 -0.028663070
#> [16,] -1.40896755 -0.002360689 -0.009622891
#> [17,] -0.59687647 -0.040091973 -0.028829438
#> [18,] -1.19567854  0.006900232 -0.020305978
#> [19,]  0.11301983 -0.239224494 -0.036518699
#> [20,]  0.70721775 -0.313096967 -0.052262068
#> [21,]  0.33006323 -0.381604820 -0.034898672
#> [22,] -0.04773094 -0.290717461 -0.036761981
#> [23,] -0.69248299 -0.261039426 -0.019910462
#> [24,] -0.07296766 -0.257178817 -0.040455484
#> [25,] -0.14157324 -0.183151857 -0.037157163
#> [26,] -0.86868253 -0.102797810 -0.019650437
#> [27,] -0.35368342 -0.110357357 -0.037775382
#> [28,] -0.18400075 -0.093266326 -0.046006746
#> [29,]  0.45593787 -0.141040453 -0.061196700
#> [30,]  0.15105230 -0.166931924 -0.050091475
#> [31,]  0.16518530 -0.242418861 -0.039776081
#> [32,] -0.61787018 -0.125817217 -0.028072443
#> [33,]  0.16242252 -0.302347456 -0.033876447
#> [34,]  0.18881026 -0.301491932 -0.038044094
#> [35,]  0.21991823 -0.210381028 -0.050205286
#> [36,] -1.13519491  0.029154617 -0.039963558
#> [37,] -1.06273340  0.107211858 -0.045158119
#> [38,] -1.25749311  0.028337402 -0.041999193
#> [39,] -1.51190638  0.002016648 -0.033316851
#> [40,] -1.51382233 -0.078979944 -0.031928841
#> [41,] -1.06637602 -0.132949508 -0.031848531
#> [42,] -0.67743132 -0.157455167 -0.044817995
#> [43,] -1.08077187 -0.134173404 -0.028863634
#> [44,] -2.00677838 -0.110081698 -0.004091785
#> [45,] -0.75051503 -0.229039489 -0.024138728
#> [46,] -1.19173718 -0.101140339 -0.032245003
#> [47,] -0.95093197 -0.037690282 -0.047994238
#> [48,] -0.15548767  0.001540447 -0.073736653
#> [49,]  0.43575002 -0.073300752 -0.082652175
#> [50,]  0.56381759 -0.014684008 -0.079868644
#> 
#> $psis$s000
#>       (Intercept) educ age
#>  [1,]           0    0   0
#>  [2,]           0    0   0
#>  [3,]           0    0   0
#>  [4,]           0    0   0
#>  [5,]           0    0   0
#>  [6,]           0    0   0
#>  [7,]           0    0   0
#>  [8,]           0    0   0
#>  [9,]           0    0   0
#> [10,]           0    0   0
#> [11,]           0    0   0
#> [12,]           0    0   0
#> [13,]           0    0   0
#> [14,]           0    0   0
#> [15,]           0    0   0
#> [16,]           0    0   0
#> [17,]           0    0   0
#> [18,]           0    0   0
#> [19,]           0    0   0
#> [20,]           0    0   0
#> [21,]           0    0   0
#> [22,]           0    0   0
#> [23,]           0    0   0
#> [24,]           0    0   0
#> [25,]           0    0   0
#> [26,]           0    0   0
#> [27,]           0    0   0
#> [28,]           0    0   0
#> [29,]           0    0   0
#> [30,]           0    0   0
#> [31,]           0    0   0
#> [32,]           0    0   0
#> [33,]           0    0   0
#> [34,]           0    0   0
#> [35,]           0    0   0
#> [36,]           0    0   0
#> [37,]           0    0   0
#> [38,]           0    0   0
#> [39,]           0    0   0
#> [40,]           0    0   0
#> [41,]           0    0   0
#> [42,]           0    0   0
#> [43,]           0    0   0
#> [44,]           0    0   0
#> [45,]           0    0   0
#> [46,]           0    0   0
#> [47,]           0    0   0
#> [48,]           0    0   0
#> [49,]           0    0   0
#> [50,]           0    0   0
#> 
#> 
#> $s.tab
#>             s111       s100      s000
#>  [1,] 0.09333333 0.61333333 0.2933333
#>  [2,] 0.05333333 0.65333333 0.2933333
#>  [3,] 0.05333333 0.59466667 0.3520000
#>  [4,] 0.05333333 0.56800000 0.3786667
#>  [5,] 0.05333333 0.51733333 0.4293333
#>  [6,] 0.05333333 0.47466667 0.4720000
#>  [7,] 0.05333333 0.42400000 0.5226667
#>  [8,] 0.05333333 0.32800000 0.6186667
#>  [9,] 0.05333333 0.26666667 0.6800000
#> [10,] 0.05333333 0.19466667 0.7520000
#> [11,] 0.05333333 0.18133333 0.7653333
#> [12,] 0.05333333 0.16000000 0.7866667
#> [13,] 0.05333333 0.14933333 0.7973333
#> [14,] 0.05333333 0.16533333 0.7813333
#> [15,] 0.05333333 0.16000000 0.7866667
#> [16,] 0.05333333 0.14933333 0.7973333
#> [17,] 0.05333333 0.13066667 0.8160000
#> [18,] 0.05333333 0.12533333 0.8213333
#> [19,] 0.05333333 0.11733333 0.8293333
#> [20,] 0.05333333 0.10133333 0.8453333
#> [21,] 0.05333333 0.10666667 0.8400000
#> [22,] 0.05600000 0.10133333 0.8426667
#> [23,] 0.05600000 0.10133333 0.8426667
#> [24,] 0.05333333 0.10400000 0.8426667
#> [25,] 0.05600000 0.11733333 0.8266667
#> [26,] 0.05600000 0.11466667 0.8293333
#> [27,] 0.05600000 0.12533333 0.8186667
#> [28,] 0.06133333 0.12266667 0.8160000
#> [29,] 0.05600000 0.09066667 0.8533333
#> [30,] 0.05333333 0.11466667 0.8320000
#> [31,] 0.05866667 0.11466667 0.8266667
#> [32,] 0.05333333 0.11733333 0.8293333
#> [33,] 0.06133333 0.08800000 0.8506667
#> [34,] 0.06133333 0.10133333 0.8373333
#> [35,] 0.06666667 0.10133333 0.8320000
#> [36,] 0.06133333 0.07733333 0.8613333
#> [37,] 0.05866667 0.11733333 0.8240000
#> [38,] 0.05866667 0.12000000 0.8213333
#> [39,] 0.06933333 0.08000000 0.8506667
#> [40,] 0.07466667 0.05333333 0.8720000
#> [41,] 0.08000000 0.04000000 0.8800000
#> [42,] 0.06666667 0.06666667 0.8666667
#> [43,] 0.06666667 0.07200000 0.8613333
#> [44,] 0.07200000 0.06400000 0.8640000
#> [45,] 0.08000000 0.04800000 0.8720000
#> [46,] 0.06933333 0.05600000 0.8746667
#> [47,] 0.07466667 0.04266667 0.8826667
#> [48,] 0.06666667 0.08000000 0.8533333
#> [49,] 0.06666667 0.09333333 0.8400000
#> [50,] 0.06666667 0.08533333 0.8480000
#> 
#> $delta.1
#>  [1]  0.698319328  0.233802817  0.154225352  0.251408451  0.328169014
#>  [6]  0.420422535  0.373239437  0.004225352 -0.190140845  0.127464789
#> [11]  0.071830986  0.186619718 -0.269718310 -0.240845070  0.231690141
#> [16]  0.411971831  0.818309859 -0.654929577  0.244366197 -0.316197183
#> [21]  0.331690141  0.045197740  0.225181598  0.595774648  0.350686037
#> [26]  0.594430993  0.400726392  0.613512846  0.515738499  0.270422535
#> [31]  0.005408190 -0.283098592  0.125617589 -0.003705534  0.268571429
#> [36] -0.107707510  0.050991501  0.389132114  0.215230328  0.067414574
#> [41]  0.466666667  0.000000000  0.251428571  0.222222222  0.101449275
#> [46]  0.575710822  0.289213668  0.742857143  0.437142857  0.314285714
#> 
#> $delta.2
#>  [1]  0.66419619  0.19118488  0.32349365  0.58479436  0.11465438  0.19484369
#>  [7]  0.59464668 -0.18763856 -0.09396083  0.04072558 -0.01761476  0.09543704
#> [13] -0.47207433 -0.32096438  0.23515557  0.44902715  0.69550700 -0.61109934
#> [19]  0.29156778 -0.31093893  0.31254138 -0.05960177  0.11975199  0.50636976
#> [25]  0.49102045  0.45251868  0.46022338  0.56250019  0.39795722  0.16104217
#> [31] -0.21833151 -0.35782657  0.10862501 -0.02622874  0.24599800 -0.04088311
#> [37] -0.11031763  0.30612624  0.29735842  0.11921913  0.38739048  0.18338280
#> [43]  0.25457316  0.20821511 -0.04280595  0.36874537  0.29757897  0.55073578
#> [49]  0.43879792  0.39496059
#>