Run Gibbs sampler without covariates

Run Gibbs sampler without covariates

Usage

prepost_gibbs_nocovar(
  formula,
  data,
  prepost,
  moderator,
  iter = 1000,
  thin = 1,
  burn = 0,
  monotonicity = TRUE,
  stable = TRUE,
  priors,
  predictive = FALSE
)

Arguments

formula

A formula with syntax y ~ t, where y is the name of the outcome variable and t is the name of the treatment.

data

A data frame containin the variables in the formula.

prepost

A one-sided formula with syntax ~ z, where z is the indicator variable for whether the moderator was measured pre- or post-treatment.

moderator

A formuala with syntax ~ d, where d is the moderator variable for the CATE.

iter

Numeric, number of iterations for the Gibbs

thin

Numeric, thinning parameter for the Gibbs

burn

Numeric, burn in rate for the Gibbs

monotonicity

A logical signifying whether Gibbs assumes monotonicity.

stable

A logical signifying whether Gibbs assumes stability.

priors

A list object containing the priors for the Gibbs sampler. Priors include beta.precision, psi.precision, alpha, y.alpha, and y.beta.

predictive

A logical indicator for whether to return prior predictive draws (TRUE) or posterior draws (FALSE, default).

Value

A list object containing Gibbs posterior quantities of interest and parameters.

Examples

data(land_experiment)
prepost_gibbs_nocovar(
  support ~ treat_comb,
  data = land_experiment,
  moderator = ~ land_insecure,
  prepost = ~ prepost,
  iter = 50
)
#> $mu
#>         0_0_s111  1_0_s111  0_1_s111  1_1_s111     0_0_s100     1_0_s100
#>  [1,] 0.26545976 0.8861003 0.4605701 0.9403564 9.980573e-04 9.837990e-01
#>  [2,] 0.38695975 0.7266110 0.3856988 0.8672870 9.589831e-04 1.069222e-01
#>  [3,] 0.12384107 0.6357363 0.3747741 0.9533995 1.012548e-01 2.442474e-08
#>  [4,] 0.40147085 0.9266365 0.5314543 1.0000000 8.231740e-01 2.350351e-01
#>  [5,] 0.21410771 0.5020546 0.7794710 0.9987843 2.479109e-01 2.400002e-01
#>  [6,] 0.12150711 0.8482997 0.5546827 0.9998481 1.171709e-01 9.996977e-01
#>  [7,] 0.23398683 0.6453511 0.6486871 0.9998998 1.000000e+00 9.981910e-01
#>  [8,] 0.31664244 0.8406047 0.5523514 0.9948552 2.034850e-05 9.999956e-01
#>  [9,] 0.20199641 0.9503121 0.7005112 0.9648504 6.105664e-01 8.384760e-01
#> [10,] 0.62010620 0.7389990 0.2419975 0.9465190 4.198325e-06 2.403918e-04
#> [11,] 0.16607896 0.9098806 0.5662852 0.9176919 1.880482e-01 8.455073e-09
#> [12,] 0.69803666 0.8493498 0.2071020 1.0000000 9.009955e-01 9.992021e-01
#> [13,] 0.17068777 0.5633276 0.8528246 0.9998515 2.416176e-04 2.397175e-01
#> [14,] 0.56039237 0.8995280 0.3186790 0.9994995 4.095113e-04 3.687620e-03
#> [15,] 0.40367405 0.9153419 0.7091893 0.8353676 9.541541e-01 4.472242e-01
#> [16,] 0.56664968 0.8775905 0.3793122 0.9829919 3.295557e-02 9.975984e-01
#> [17,] 0.61837729 0.7902128 0.5184542 0.9779346 8.545044e-01 9.992175e-01
#> [18,] 0.14144287 0.9152376 0.6093261 0.8208263 7.775985e-09 4.095290e-01
#> [19,] 0.42531873 0.7635679 0.4437349 0.8995399 5.641933e-04 3.379747e-06
#> [20,] 0.33699944 0.7994096 0.4277309 0.9727310 5.105672e-01 1.742551e-02
#> [21,] 0.42920143 0.8511382 0.6693029 0.9997445 4.344036e-01 1.469105e-01
#> [22,] 0.52047204 0.8245460 0.4083795 0.9842599 2.841919e-01 9.172109e-01
#> [23,] 0.52622409 0.7513593 0.7468844 1.0000000 6.261252e-01 9.588591e-01
#> [24,] 0.19155464 0.7730742 0.2698372 0.9956876 1.559000e-01 9.993112e-01
#> [25,] 0.72732786 0.9298617 0.3467241 0.9971751 1.935415e-06 9.863149e-01
#> [26,] 0.06936517 0.7723045 0.5729815 0.9999679 1.032262e-01 9.999996e-01
#> [27,] 0.55402043 0.6390398 0.2654724 0.9914403 8.819972e-02 7.319072e-01
#> [28,] 0.31992568 0.8731474 0.6668624 0.9370218 5.439427e-01 7.855599e-01
#> [29,] 0.37916400 0.9285139 0.3475264 0.9999568 3.130772e-01 7.183074e-01
#> [30,] 0.15363304 0.6182488 0.7017020 0.9999954 3.283290e-06 4.876210e-01
#> [31,] 0.22472953 0.9664175 0.2495204 0.9987758 6.283944e-05 8.058631e-01
#> [32,] 0.25947393 0.6819406 0.6612764 0.9958613 8.646172e-01 9.990857e-01
#> [33,] 0.56753082 0.6145569 0.6341170 0.9992633 9.926318e-01 5.222733e-01
#> [34,] 0.20534150 0.9381480 0.6645967 0.9985556 3.424622e-02 6.557343e-01
#> [35,] 0.71098232 0.7455668 0.6939515 0.9783603 1.168995e-02 1.000000e+00
#> [36,] 0.56745596 0.7219166 0.2886271 0.9684186 3.951311e-05 9.993258e-01
#> [37,] 0.54330234 0.6193193 0.5343873 0.8765548 4.867525e-02 9.612207e-01
#> [38,] 0.57734848 0.8016571 0.5970516 0.9827075 2.886576e-01 8.523149e-01
#> [39,] 0.13366037 0.7809246 0.2243318 0.9331283 1.216836e-01 9.999997e-01
#> [40,] 0.25647051 0.7694878 0.4444465 0.9064589 3.728031e-03 8.246369e-01
#> [41,] 0.13202110 0.7023853 0.6992695 0.8607982 7.763406e-02 3.017838e-04
#> [42,] 0.09822506 0.7830706 0.4516593 0.7963781 9.999363e-01 9.932561e-01
#> [43,] 0.33458931 0.7983244 0.1181678 0.9688758 9.827494e-01 2.595833e-02
#> [44,] 0.49028535 0.8848900 0.1959401 0.9673246 2.066095e-04 6.101250e-04
#> [45,] 0.53944978 0.8294450 0.6463091 0.9938488 8.651565e-06 7.999595e-01
#> [46,] 0.21544133 0.6794633 0.2427270 0.9321414 9.999748e-01 9.975662e-01
#> [47,] 0.57592308 0.6550406 0.3411180 0.8660219 4.501476e-02 1.172675e-02
#> [48,] 0.17059011 0.8428276 0.6253291 0.9211732 1.436914e-02 9.144544e-01
#> [49,] 0.55472850 0.7008558 0.6165234 0.8763502 8.105037e-01 1.960705e-01
#> [50,] 0.24940226 0.9250172 0.7644916 0.9970212 9.955597e-01 3.204693e-01
#>           0_1_s100     1_1_s100  0_0_s000  1_0_s000  0_1_s000  1_1_s000
#>  [1,] 5.632014e-01 2.190527e-01 0.9182560 0.8479071 0.8572487 0.7774204
#>  [2,] 1.257379e-06 9.837610e-01 0.9336154 0.8549985 0.9089162 0.6887129
#>  [3,] 4.585156e-01 5.621618e-05 0.9184039 0.8910188 0.8711393 0.8084534
#>  [4,] 2.820829e-01 4.047631e-02 0.8382306 0.8897875 0.8542351 0.7575839
#>  [5,] 6.630731e-04 9.710247e-06 0.9281514 0.8412488 0.8718417 0.7975691
#>  [6,] 2.530445e-01 9.649394e-03 0.8135634 0.8531508 0.8055736 0.8523939
#>  [7,] 3.833459e-07 7.300431e-03 0.9503148 0.8477834 0.7756018 0.8576002
#>  [8,] 2.233751e-04 2.340118e-02 0.9332306 0.8857717 0.7721333 0.8131409
#>  [9,] 9.999997e-01 3.074937e-01 0.8961136 0.9178630 0.8241766 0.8577969
#> [10,] 8.065239e-01 3.602518e-02 0.9234270 0.9353442 0.8167716 0.7682100
#> [11,] 9.996079e-01 5.326227e-03 0.8806178 0.8139432 0.8936407 0.7812077
#> [12,] 9.965642e-01 7.606338e-02 0.8585324 0.9037066 0.8238371 0.8079662
#> [13,] 9.884150e-01 6.356176e-01 0.9229818 0.8330473 0.8399879 0.7469164
#> [14,] 7.258407e-02 1.872351e-01 0.9247422 0.9006774 0.8208176 0.8228399
#> [15,] 9.990859e-01 1.486793e-01 0.8420262 0.8831553 0.9196742 0.7911308
#> [16,] 6.074156e-01 9.981512e-01 0.9098755 0.8784668 0.8930882 0.7966956
#> [17,] 2.689075e-01 9.986078e-01 0.8937864 0.8683965 0.8886204 0.7214773
#> [18,] 1.467621e-01 9.997486e-01 0.9370417 0.8887537 0.8208519 0.8291719
#> [19,] 9.999980e-01 8.772496e-01 0.8982199 0.8954905 0.8633684 0.7250094
#> [20,] 9.122916e-01 3.011172e-01 0.9209500 0.9105441 0.6656620 0.8178781
#> [21,] 2.662759e-06 3.049546e-01 0.9529421 0.8686595 0.8288820 0.7930798
#> [22,] 1.000000e+00 3.182396e-01 0.9101095 0.9119482 0.7911245 0.7609182
#> [23,] 8.639864e-03 7.566071e-01 0.8637262 0.8710458 0.8802959 0.7482451
#> [24,] 1.110564e-05 6.193956e-01 0.9030673 0.8820508 0.8789689 0.8088976
#> [25,] 7.891812e-03 1.746538e-01 0.8962971 0.8550465 0.8080393 0.7518617
#> [26,] 1.864810e-02 1.207743e-02 0.9718270 0.8464502 0.8911498 0.7641176
#> [27,] 1.340633e-01 3.394264e-02 0.9435040 0.8565011 0.9549624 0.7963908
#> [28,] 1.946469e-01 5.188750e-03 0.9453781 0.8160351 0.9642446 0.7593281
#> [29,] 4.135843e-02 1.913158e-01 0.9306489 0.8617778 0.9443930 0.7352722
#> [30,] 7.699448e-03 9.139489e-05 0.9186588 0.9021489 0.8402438 0.8312019
#> [31,] 9.958186e-01 6.141157e-02 0.8655206 0.8657934 0.8377824 0.7759123
#> [32,] 1.000000e+00 5.773844e-01 0.9262813 0.9329654 0.8385529 0.8058232
#> [33,] 1.000000e+00 5.151088e-01 0.8573888 0.8690442 0.7522409 0.7748908
#> [34,] 9.999907e-01 6.102487e-01 0.9365142 0.8824248 0.8682420 0.7723225
#> [35,] 2.265091e-01 9.327610e-01 0.9222421 0.8612964 0.7863521 0.7965588
#> [36,] 3.818926e-01 9.999960e-01 0.8530454 0.8758836 0.7736389 0.7855373
#> [37,] 6.753358e-01 9.434521e-01 0.9286278 0.8272194 0.9325652 0.8013987
#> [38,] 9.999875e-01 9.888499e-01 0.8603674 0.9198417 0.6603062 0.7580122
#> [39,] 9.839237e-01 9.999313e-01 0.9295255 0.8647074 0.8440015 0.8511525
#> [40,] 9.607016e-01 9.917558e-01 0.8428603 0.8529322 0.8522906 0.7992557
#> [41,] 1.771611e-04 9.999979e-01 0.8725378 0.8774558 0.7650832 0.7389420
#> [42,] 5.131016e-02 9.646178e-01 0.8971662 0.8345526 0.8553760 0.8474407
#> [43,] 1.257299e-04 9.999985e-01 0.8287696 0.8845845 0.8288744 0.7659484
#> [44,] 9.997450e-01 9.921868e-01 0.8858439 0.8463296 0.8585562 0.7815511
#> [45,] 9.438560e-01 4.179291e-01 0.9112640 0.8069229 0.7394514 0.8485042
#> [46,] 1.671809e-03 8.425924e-04 0.8392477 0.8674550 0.8399991 0.8341047
#> [47,] 2.156046e-02 2.775102e-01 0.9290975 0.8588999 0.8327411 0.8074301
#> [48,] 5.395299e-01 6.324605e-01 0.9385663 0.9315347 0.8478532 0.8113843
#> [49,] 8.922365e-01 5.659966e-04 0.9139122 0.8729566 0.9067384 0.7677053
#> [50,] 9.470336e-01 9.988841e-01 0.7741925 0.8442440 0.8490298 0.8263637
#> 
#> $psis
#>             s111         s100      s000
#>  [1,] 0.08317584 2.563187e-03 0.9142610
#>  [2,] 0.09931538 6.798888e-03 0.8938857
#>  [3,] 0.09820408 8.857408e-03 0.8929385
#>  [4,] 0.09538971 8.283696e-03 0.8963266
#>  [5,] 0.09701934 1.834897e-02 0.8846317
#>  [6,] 0.08850052 1.737473e-02 0.8941247
#>  [7,] 0.07311797 8.437421e-03 0.9184446
#>  [8,] 0.08987243 1.549871e-02 0.8946289
#>  [9,] 0.07586830 3.954296e-03 0.9201774
#> [10,] 0.08743814 4.308588e-03 0.9082533
#> [11,] 0.09582706 4.895371e-03 0.8992776
#> [12,] 0.09699266 7.163713e-03 0.8958436
#> [13,] 0.09588536 2.273734e-03 0.9018409
#> [14,] 0.08425182 3.657887e-04 0.9153824
#> [15,] 0.10367916 5.072404e-03 0.8912484
#> [16,] 0.08247057 4.418971e-03 0.9131105
#> [17,] 0.09281426 3.031818e-03 0.9041539
#> [18,] 0.09467363 3.976390e-03 0.9013500
#> [19,] 0.09898127 4.899810e-03 0.8961189
#> [20,] 0.07877555 1.107375e-02 0.9101507
#> [21,] 0.08981219 1.501432e-02 0.8951735
#> [22,] 0.11077683 1.483498e-02 0.8743882
#> [23,] 0.07631590 2.803263e-02 0.8956515
#> [24,] 0.10116208 2.311488e-02 0.8757230
#> [25,] 0.09121439 2.652865e-02 0.8822570
#> [26,] 0.07925243 5.956920e-02 0.8611784
#> [27,] 0.07920282 5.749136e-02 0.8633058
#> [28,] 0.08232435 4.074153e-02 0.8769341
#> [29,] 0.11172247 2.686836e-02 0.8614092
#> [30,] 0.06009525 3.095481e-02 0.9089499
#> [31,] 0.08668315 1.340756e-02 0.8999093
#> [32,] 0.05781669 8.945752e-03 0.9332376
#> [33,] 0.11004773 1.984963e-02 0.8701026
#> [34,] 0.10484989 3.282023e-02 0.8623299
#> [35,] 0.06397661 4.791348e-02 0.8881099
#> [36,] 0.10642550 3.212936e-02 0.8614451
#> [37,] 0.09120804 4.085779e-02 0.8679342
#> [38,] 0.08631567 3.269492e-02 0.8809894
#> [39,] 0.06376098 1.030980e-02 0.9259292
#> [40,] 0.07653361 8.445927e-03 0.9150205
#> [41,] 0.10530109 2.342502e-03 0.8923564
#> [42,] 0.08691081 3.130320e-03 0.9099589
#> [43,] 0.10827345 1.267077e-04 0.8915998
#> [44,] 0.11063598 1.686652e-03 0.8876774
#> [45,] 0.10015794 6.241244e-10 0.8998421
#> [46,] 0.11364032 8.368605e-05 0.8862760
#> [47,] 0.07481612 1.144352e-05 0.9251724
#> [48,] 0.10074912 1.596944e-04 0.8990912
#> [49,] 0.09163858 1.572426e-03 0.9067890
#> [50,] 0.11151403 2.109642e-03 0.8863763
#> 
#> $s.tab
#>             s111        s100      s000
#>  [1,] 0.09333333 0.008000000 0.8986667
#>  [2,] 0.09333333 0.002666667 0.9040000
#>  [3,] 0.09333333 0.008000000 0.8986667
#>  [4,] 0.09066667 0.010666667 0.8986667
#>  [5,] 0.09066667 0.018666667 0.8906667
#>  [6,] 0.09066667 0.010666667 0.8986667
#>  [7,] 0.09066667 0.008000000 0.9013333
#>  [8,] 0.09066667 0.008000000 0.9013333
#>  [9,] 0.09066667 0.005333333 0.9040000
#> [10,] 0.09333333 0.005333333 0.9013333
#> [11,] 0.09333333 0.005333333 0.9013333
#> [12,] 0.09066667 0.010666667 0.8986667
#> [13,] 0.09066667 0.002666667 0.9066667
#> [14,] 0.09066667 0.005333333 0.9040000
#> [15,] 0.09333333 0.000000000 0.9066667
#> [16,] 0.09333333 0.002666667 0.9040000
#> [17,] 0.08800000 0.005333333 0.9066667
#> [18,] 0.09066667 0.005333333 0.9040000
#> [19,] 0.08800000 0.005333333 0.9066667
#> [20,] 0.09066667 0.013333333 0.8960000
#> [21,] 0.08800000 0.018666667 0.8933333
#> [22,] 0.09066667 0.026666667 0.8826667
#> [23,] 0.08800000 0.018666667 0.8933333
#> [24,] 0.08533333 0.026666667 0.8880000
#> [25,] 0.08800000 0.024000000 0.8880000
#> [26,] 0.09066667 0.040000000 0.8693333
#> [27,] 0.09066667 0.050666667 0.8586667
#> [28,] 0.09066667 0.045333333 0.8640000
#> [29,] 0.09066667 0.042666667 0.8666667
#> [30,] 0.09066667 0.024000000 0.8853333
#> [31,] 0.09066667 0.008000000 0.9013333
#> [32,] 0.08800000 0.010666667 0.9013333
#> [33,] 0.08533333 0.021333333 0.8933333
#> [34,] 0.08800000 0.024000000 0.8880000
#> [35,] 0.08533333 0.034666667 0.8800000
#> [36,] 0.08000000 0.034666667 0.8853333
#> [37,] 0.08533333 0.040000000 0.8746667
#> [38,] 0.08800000 0.026666667 0.8853333
#> [39,] 0.08533333 0.016000000 0.8986667
#> [40,] 0.09333333 0.008000000 0.8986667
#> [41,] 0.09066667 0.002666667 0.9066667
#> [42,] 0.09333333 0.002666667 0.9040000
#> [43,] 0.09066667 0.002666667 0.9066667
#> [44,] 0.09333333 0.000000000 0.9066667
#> [45,] 0.09333333 0.000000000 0.9066667
#> [46,] 0.09333333 0.000000000 0.9066667
#> [47,] 0.09333333 0.000000000 0.9066667
#> [48,] 0.09333333 0.000000000 0.9066667
#> [49,] 0.09333333 0.000000000 0.9066667
#> [50,] 0.09333333 0.002666667 0.9040000
#> 
#> $delta.1
#>  [1] 0.6025210 0.6995798 0.6369748 0.5144040 0.3056753 0.5264792 0.3148180
#>  [8] 0.4795584 0.4059859 0.6273109 0.5718487 0.8644989 0.2879938 0.7234777
#> [15] 0.2516807 0.6831933 0.5879851 0.1617216 0.4444444 0.6355874 0.3107390
#> [22] 0.6586165 0.3878256 0.6649964 0.6948432 0.6142832 0.7936864 0.4554080
#> [29] 0.7348629 0.3645851 0.7057098 0.4944179 0.3262573 0.4885699 0.4774964
#> [36] 0.7753623 0.5337099 0.3423711 0.6854045 0.5768908 0.2852337 0.3714286
#> [43] 0.7822149 0.7277311 0.1974790 0.6491597 0.2285714 0.2991597 0.3008403
#> [50] 0.3584034
#> 
#> $delta.2
#>  [1] 0.5603535 0.6927033 0.6451985 0.5665243 0.2920900 0.4038770 0.2698942
#>  [8] 0.4017999 0.2338259 0.7564916 0.4686142 0.8159455 0.2407517 0.6787533
#> [15] 0.2588067 0.6977262 0.6236263 0.1994703 0.5940791 0.4019605 0.3606228
#> [22] 0.6169566 0.3584590 0.7781914 0.7001208 0.5462252 0.8808901 0.4743896
#> [29] 0.8506898 0.3072881 0.8239344 0.3710164 0.3538167 0.4406512 0.2385727
#> [36] 0.6460965 0.4553829 0.2918446 0.7015480 0.5142782 0.1849837 0.3494958
#> [43] 0.9134830 0.8482580 0.2384869 0.6953082 0.5502114 0.3322900 0.4001628
#> [50] 0.2550188
#> 
#> $loglike
#>  [1] 279.1851 282.9807 279.9751 279.0636 282.2646 282.4640 284.2438 279.4488
#>  [9] 283.4940 281.6907 281.9222 280.3550 284.4148 282.1858 282.3042 279.8755
#> [17] 280.3595 281.6967 279.0994 282.7396 279.6937 279.9020 278.6587 278.8682
#> [25] 281.5482 284.9137 284.4510 286.7399 283.4021 284.0434 281.1932 283.9315
#> [33] 280.5197 280.4500 280.0763 281.0671 280.8210 283.6936 283.0973 278.7983
#> [41] 281.6617 282.2489 281.6991 280.7762 284.8978 281.0114 280.7746 282.2424
#> [49] 280.1402 285.7813
#> 
#> $logpost
#>  [1] -280.1645 -275.0406 -267.4364 -270.2956 -275.0142 -278.1046 -253.5943
#>  [8] -262.3711 -278.0633 -273.1875 -267.3070       Inf -279.0414 -274.2703
#> [15] -284.7876 -277.6109 -278.6468 -269.2438 -263.9983 -288.5137 -274.3176
#> [22] -271.7868 -270.1102 -272.6147 -273.4278 -271.0759 -289.3037 -291.9495
#> [29] -284.1970 -267.0467 -274.6235 -271.2789 -270.8574 -276.1599 -270.8856
#> [36] -270.4466 -287.5380 -282.2315 -270.4845 -280.6702 -268.7067 -278.3420
#> [43] -267.0301 -268.8207 -270.4344 -266.9901 -279.8475 -284.2849 -282.3045
#> [50] -282.1781
#>