Run Gibbs sampler for the random moderator placement design

Run Gibbs sampler for the random moderator placement design

Usage

prepost_gibbs(
  formula,
  data,
  prepost,
  moderator,
  covariates,
  iter = 1000,
  thin = 1,
  burn = 0,
  offset = 0,
  monotonicity = TRUE,
  stable = TRUE,
  saturated = TRUE,
  priors
)

Arguments

formula

A formula with syntax y ~ t, where y is the (unquoted) name of the outcome and t is the (unquoted) name of the treatment.

data

A data.frame containing variables in the formula, moderator, and covariates arguments.

prepost

A one-sided formula with syntax ~ z, where z is the indicator variable for whether the moderator was measured pre- or post-treatment.

moderator

A one-sided formuala with syntax ~ d, where d is the (unquoted) name of the moderator variable for the CATE.

covariates

A one-sided formula with syntax ~ x1 + x2, where the right-hand side variables signify which covariates the Gibbs will use to try and narrow the bounds.

iter

Integer indicating the number of iterations for the Gibbs sampler.

thin

Integer indicating how often the Gibbs sampler should save a draw.

burn

Integer indicating how many iterations should be performed before saving draws in the Gibbs sampler.

offset

A numeric value indicating the center of the prior distribution for the covariate coefficients.

monotonicity

A logical signifying whether the model assumes monotonicity.

stable

A logical signifying whether the model assumes that the the pre vs post indicator does not affect the moderator under the control condition for treatment.

saturated

A logical indicating whether the coefficients on the covariates are allowed to vary by the principal strata.

priors

A list object containing the priors for the Gibbs sampler. Priors include beta.precision, psi.precision, alpha, y.alpha, and y.beta.

Value

A list object containing Gibbs posterior quantities of interest and parameters.

Examples

data(land_experiment)
prepost_gibbs(
  support ~ treat_comb,
  data = land_experiment,
  moderator = ~ land_insecure,
  prepost = ~ prepost,
  covariates = ~ educ + age,
  iter = 50
)
#> $betas
#>       (Intercept)        educ           age   treat_comb     prepost
#>  [1,]   30.212575 -11.8537972  2.7309365459  4.823811749  5.29340999
#>  [2,]   17.347981 -11.6205965  1.3750416798  2.376185810  6.01355064
#>  [3,]   17.063923  -8.2752723  0.7386389708 -0.668864652 -2.91078855
#>  [4,]   12.183307  -5.5925127  0.3632198760  0.454410167  1.80916241
#>  [5,]   13.688792  -4.2170128  0.2033732662 -0.861457855  0.37006097
#>  [6,]   10.922087  -2.7178304  0.0746654762 -0.175505732 -0.10901944
#>  [7,]   11.006588  -2.0502704  0.0147585403 -1.086761237 -0.51287137
#>  [8,]    8.950237  -1.5977228  0.0033455239 -0.106795931  0.37992285
#>  [9,]    7.330855  -1.0669566 -0.0254633815  0.399247503  0.30056232
#> [10,]    7.297227  -0.9357218 -0.0142156622 -0.476618773 -1.02979146
#> [11,]    4.912926  -0.6553579  0.0083146234  0.242060485 -0.12586741
#> [12,]    4.901685  -0.5969041 -0.0121492009  0.271876410 -1.02519872
#> [13,]    4.040699  -0.3545045 -0.0161197523  0.119241875 -0.54934397
#> [14,]    4.145066  -0.4392895 -0.0270533618  0.445127049  0.83497893
#> [15,]    3.599293  -0.4526549 -0.0064483567  0.512666844  0.57357107
#> [16,]    4.694732  -0.4779392 -0.0233520915  0.175285515  0.27727913
#> [17,]    3.428479  -0.3548127 -0.0059827652  0.497414121 -0.73498836
#> [18,]    3.800330  -0.2594181 -0.0192530758 -0.043071991 -0.70506497
#> [19,]    4.003899  -0.3124689 -0.0152006947  0.008227955 -0.24163257
#> [20,]    2.804720  -0.2600400 -0.0049623161  0.441984856  0.03918917
#> [21,]    2.747150  -0.3556296  0.0058852814  0.786562649  0.15980025
#> [22,]    2.985998  -0.3338013  0.0005566419  0.281011379  0.05687172
#> [23,]    2.791271  -0.3408166  0.0086912037  0.208047340 -0.36534928
#> [24,]    4.073964  -0.4512847 -0.0109389694 -0.362750012 -0.23847874
#> [25,]    4.325852  -0.3941545 -0.0291004076  0.409961672  0.54395510
#> [26,]    3.707239  -0.3387944 -0.0212457590 -0.018005373  0.73021423
#> [27,]    4.029704  -0.3263130 -0.0256834692  0.058645520 -0.54139032
#> [28,]    4.249006  -0.3347559 -0.0236237130 -0.305848090 -0.38014175
#> [29,]    4.974450  -0.3493922 -0.0403770422 -0.318752854 -0.64272830
#> [30,]    3.732185  -0.1944656 -0.0148292743 -0.381111574 -0.62270158
#> [31,]    4.846566  -0.4276565 -0.0249360265 -0.089333521 -0.52596283
#> [32,]    4.292095  -0.4004673 -0.0168863283  0.213068250 -0.27712678
#> [33,]    4.895765  -0.4489806 -0.0336228212  0.239584188 -0.40395924
#> [34,]    3.681832  -0.3828890 -0.0172059098  0.376950840  0.19585001
#> [35,]    4.188628  -0.3916016 -0.0205702133  0.046475689  0.28684475
#> [36,]    5.324942  -0.3947830 -0.0309248254 -0.361720247 -1.12355534
#> [37,]    4.944411  -0.3729182 -0.0306921257  0.150722021 -0.62928035
#> [38,]    4.486559  -0.4623836 -0.0287314194  0.554606711 -0.29102027
#> [39,]    4.364822  -0.3101202 -0.0278291419 -0.030334676 -0.06091105
#> [40,]    4.506959  -0.4318025 -0.0180695375  0.533079710 -0.31007643
#> [41,]    4.836545  -0.4973782 -0.0284422474  0.411153898  0.36543160
#> [42,]    4.614945  -0.4270191 -0.0353621807  0.246796047  0.74254676
#> [43,]    5.240896  -0.5073752 -0.0344816041  0.172709983 -0.13730172
#> [44,]    4.534536  -0.3864195 -0.0270793686  0.165161959 -0.05417443
#> [45,]    4.712136  -0.4405812 -0.0330813841  0.458692730 -0.61694161
#> [46,]    3.471315  -0.4166839 -0.0059984988  0.857158838  0.92978014
#> [47,]    3.759170  -0.3206376 -0.0063121806  0.471160481 -0.60242632
#> [48,]    3.090188  -0.4698050  0.0042258959  1.130620193 -0.20015615
#> [49,]    2.306876  -0.2739481  0.0077575500  0.708580982 -0.02784993
#> [50,]    3.509441  -0.3476037 -0.0058468577 -0.222844043 -0.88406421
#>       treat_comb:prepost        s111       s100 treat_comb:s111 treat_comb:s100
#>  [1,]        12.38508036 -16.0854471 34.8900396     -1.79562827       2.7953828
#>  [2,]        28.77298651   6.1564731 66.9314577      6.75905661      14.3515350
#>  [3,]        26.71017604   5.0895341 62.4176858     -0.09275494      20.8968924
#>  [4,]        15.27585025   0.8378360 60.7336731      3.73337903      21.2961738
#>  [5,]         9.84313207  -1.6704474 54.5631320      1.09425860      17.8944962
#>  [6,]         5.88857726  -1.7982097 51.0449948      2.09550292      16.9745082
#>  [7,]         3.60591852   0.2266369 44.0807209     -0.27072828      17.8993154
#>  [8,]         1.01684010  -2.0359767 37.0692414      2.11208211      21.6668082
#>  [9,]         0.72957869  -1.4214501 33.9543130      0.96716921      20.0359952
#> [10,]         0.90020319  -2.3939860 30.5886513      0.90565259      20.6824489
#> [11,]        -0.54519334  -1.2987377 26.8148044      0.53032047      18.6326663
#> [12,]        -0.13449216  -1.3871678 25.1130590      1.11029700      16.5151741
#> [13,]         0.01966229  -2.9061950 25.9338554      1.73804931      11.1488347
#> [14,]        -0.88912130  -1.6731489 20.5733907      1.19108026      11.1639277
#> [15,]        -0.77906435  -1.7343319 17.0809838      0.11113084      12.5230598
#> [16,]        -0.84890599  -0.5362716 19.2482317     -0.32387545      10.0894386
#> [17,]        -0.12872987  -0.2569527 17.3164174      0.89757804      13.4261073
#> [18,]         0.02809284  -2.4145838 14.5268147      1.92596028      10.3813563
#> [19,]        -0.69823663  -2.2667494 14.9392755      1.44796965      12.2762492
#> [20,]        -0.72669377  -1.1833967 13.4602731     -0.64486009      14.9684720
#> [21,]        -1.20576174  -0.8309939 15.3468947     -0.99437209      11.7998424
#> [22,]        -0.64293179  -3.3399519 17.1281254      1.60335217       7.1953350
#> [23,]        -0.13304321  -0.8635163 18.2163531      0.53298409       4.8727568
#> [24,]         0.42955891  -2.8840753 18.4868750      0.74379101       4.8801985
#> [25,]        -1.37236594  -1.2790325 12.8530781      0.26864462       8.5853668
#> [26,]        -1.13946345  -1.9378596 13.9491071      1.08065176      10.6051773
#> [27,]        -0.46392840  -3.1875464  9.8933042      3.69908846      12.7339749
#> [28,]        -0.13817933  -1.1550524  7.8714792      0.10462730      11.7705331
#> [29,]         0.31080921  -1.8624195 10.5050470      1.55705477       5.6645653
#> [30,]         0.25821147  -2.7913742  8.8717427      1.80600523       5.9997861
#> [31,]         0.20251689  -3.5938461  7.1395115      2.56021898       5.7203134
#> [32,]        -0.70491284  -2.5842031  5.9880668      1.52305017       8.3868046
#> [33,]        -0.13888481  -1.2428148  5.7400437      2.51895996       3.2404098
#> [34,]        -0.78688316  -0.5314306  1.9704547     -0.77282832       4.3935361
#> [35,]        -0.84177185  -1.5809878  2.0634683      2.03331287       3.9465056
#> [36,]         0.41488758  -3.1805812  0.4524743      3.10301080       5.7966402
#> [37,]        -0.56582182  -1.1518807 -0.2815877      0.63092544       5.0518705
#> [38,]        -0.68149696  -0.9446442  0.3661174     -0.89006578       4.2246814
#> [39,]        -0.46278836  -1.6685370 -1.0881035      0.96888117       5.5890004
#> [40,]        -0.93161729  -3.0423034  0.3737793      1.56632219       5.5037868
#> [41,]        -1.39521079  -2.0732773  2.0801083      1.48960692       2.8872792
#> [42,]        -1.00882729  -0.1659696  2.7554272     -0.04853148       2.9687928
#> [43,]        -0.45263780  -2.1811078  3.0003752      2.98964639       1.1511771
#> [44,]        -0.25057656  -1.8456675  2.5776401      2.08528871       2.4649329
#> [45,]        -0.28530712  -1.6430893  0.1855946      2.20070839       4.2198279
#> [46,]        -2.43454411  -1.2166178  1.3038905     -0.63598625       2.8369106
#> [47,]        -1.00295627  -2.7132943  4.2654718      1.05444450       0.8263301
#> [48,]        -0.94581353  -1.5824133  3.8272797      3.30484152       3.4094391
#> [49,]        -1.14871852  -0.1807384  3.6623857      0.94106057       2.2826558
#> [50,]         1.02159834  -2.7651585  2.2549343      1.84652234       3.7990027
#>       prepost:s111 prepost:s100 treat_comb:prepost:s111 treat_comb:prepost:s100
#>  [1,]  -9.08902544    2.5065562              -1.9102114              -1.6357955
#>  [2,]  -6.04116931   -4.5800302               3.2972060              -7.8315901
#>  [3,]  -1.83960951    9.9069841               2.5480292             -10.0029850
#>  [4,]   0.76515878    8.5341838              -1.7267658              -4.0362529
#>  [5,]   3.98573927    7.3881394               0.2924717              -3.8505013
#>  [6,]   1.99432789    8.6845133              -0.7771206              -1.1736642
#>  [7,]  -2.05745721    9.0644806              -6.4011523               2.0080721
#>  [8,]   0.80188072   13.9227472              -6.4055757              -0.5635995
#>  [9,]  -1.51562755   12.4858726              -7.1106378               4.6673503
#> [10,]  -0.84422276   13.2238419              -0.8068188               6.5682080
#> [11,]   0.57702465   16.2733360              -5.5206109               5.6827149
#> [12,]   0.44903866   15.9719396              -2.9385794               8.0775324
#> [13,]   1.42879060   11.7598871              -1.9865231               8.4678067
#> [14,]  -2.25551543    9.3170962              -0.4088152               7.5948510
#> [15,]  -0.73866062   12.4435483              -2.9354140               7.9100637
#> [16,]   0.04766461    5.9542653              -2.1396561              10.6128300
#> [17,]  -0.07435571    7.5887424              -2.1384005               5.3791589
#> [18,]   0.94717612   12.4675361              -3.7307751              10.3642018
#> [19,]   0.86540919    8.5894971              -6.3453442               8.4548159
#> [20,]  -0.73225538    6.7724095              -6.9372600               7.2637186
#> [21,]   0.99141474    7.2876928              -3.6139709               1.3687238
#> [22,]   0.86498527    9.1330914              -1.9725577               1.7890724
#> [23,]   0.94614757   10.1367176              -1.0640929               4.0307486
#> [24,]   2.92521809    3.4780080              -1.9440249               6.8571470
#> [25,]  -0.90721860    4.9293471               2.6082751               8.2009915
#> [26,]  -0.56022435    4.4157635               0.8284288               3.2464224
#> [27,]   3.11704872    7.6102408               0.5458453               1.8141630
#> [28,]  -0.19028395    8.8938214              -0.4457905               2.0348497
#> [29,]   1.45686097    4.5905190              -3.9542227               4.2876636
#> [30,]   1.13208443    4.8415439              -2.2252209               3.3745594
#> [31,]   1.30319413    4.2380672              -1.8316654               0.2516719
#> [32,]   1.21546897    3.6964359              -3.7863391               0.2422055
#> [33,]   0.83148810    3.2634909              -1.0104577               4.0925002
#> [34,]  -0.42768511    7.5488986               3.6093205              -1.5474775
#> [35,]   0.02999038    0.3913469               3.5460432               3.7848681
#> [36,]   2.75499679    3.5256066              -0.2035966              -0.2483785
#> [37,]   0.84763188   -0.6312945               3.1880497               5.2939406
#> [38,]  -0.19976060   -0.8324727               3.6544257               4.6877014
#> [39,]   0.85203154    0.5797567              -1.6306955               1.4190388
#> [40,]   0.90376297   -0.4233247              -0.2661265               0.5591580
#> [41,]   0.04960359   -0.6288505              -0.8975335               3.8465255
#> [42,]  -2.30668322    0.3872884               2.5790673               2.0355833
#> [43,]   0.77114443    1.9363781              -3.0080613               1.0464749
#> [44,]  -0.61724503   -0.3787547              -0.4978001               3.9426253
#> [45,]   1.16810714    4.8876608               1.6015822              -3.5990264
#> [46,]   0.69610217    2.0866492               2.5055598              -1.7740828
#> [47,]   2.14275554    3.9417728              -0.6139898              -3.5016447
#> [48,]   0.54297338    3.1377323              -1.1425364              -4.8285481
#> [49,]   0.39578283    2.1694092              -0.2807254               0.8186198
#> [50,]   0.47377589    4.6096484              -0.3773665              -1.0267630
#> 
#> $psis
#> $psis$s111
#>       (Intercept)          educ         age
#>  [1,]  -10.105394 -12.256262889 -2.60521207
#>  [2,]  -16.325836 -10.183464326 -2.02589106
#>  [3,]  -19.803014  -8.209267560 -1.67744206
#>  [4,]  -20.013498  -7.463966892 -1.36330815
#>  [5,]  -17.587259  -6.895760593 -1.15466787
#>  [6,]  -18.610069  -6.201205263 -0.90459214
#>  [7,]  -19.329208  -5.496692955 -0.71912031
#>  [8,]  -17.701598  -5.305527140 -0.55344911
#>  [9,]  -19.293968  -4.606317402 -0.36022775
#> [10,]  -15.294976  -4.755113866 -0.29588882
#> [11,]  -15.797084  -4.004791102 -0.19816940
#> [12,]  -12.355360  -4.049429067 -0.16425935
#> [13,]  -14.002349  -3.623110012 -0.06220627
#> [14,]  -19.142667  -2.625839880  0.04555993
#> [15,]  -18.478821  -2.320909190  0.07761438
#> [16,]  -19.540601  -2.128627443  0.15049983
#> [17,]  -18.556733  -1.814596711  0.15553779
#> [18,]  -15.500877  -1.504175072  0.10797882
#> [19,]  -13.602548  -1.469152059  0.12648841
#> [20,]  -14.302768  -1.206809069  0.16860515
#> [21,]  -12.299225  -1.391049074  0.19379851
#> [22,]  -10.560982  -1.073460760  0.13909171
#> [23,]   -8.465825  -0.988820745  0.11400070
#> [24,]   -7.913076  -0.642180681  0.10117375
#> [25,]   -7.839936  -0.484262163  0.10023441
#> [26,]   -7.622826  -0.426068975  0.10307026
#> [27,]   -7.278693  -0.271652505  0.09103297
#> [28,]   -6.684070  -0.223556029  0.08211267
#> [29,]   -6.255361  -0.184722456  0.07539422
#> [30,]   -6.804764   0.072830310  0.07833816
#> [31,]   -4.648826   0.042312260  0.03415097
#> [32,]   -3.381602  -0.216536542  0.03296244
#> [33,]   -3.958780  -0.121135648  0.03398095
#> [34,]   -3.483508  -0.100245245  0.02332267
#> [35,]   -3.511647  -0.034093286  0.01703376
#> [36,]   -2.731719  -0.081011303  0.01126985
#> [37,]   -4.407543   0.068783145  0.03412813
#> [38,]   -5.053451   0.167632691  0.03895857
#> [39,]   -3.983048   0.037658323  0.02428336
#> [40,]   -3.894394   0.039493014  0.03363367
#> [41,]   -4.718815   0.249970415  0.03842147
#> [42,]   -4.550900   0.205779079  0.03532320
#> [43,]   -3.483999  -0.004229671  0.02043677
#> [44,]   -3.528087   0.003051194  0.03259845
#> [45,]   -4.897411   0.166237230  0.04568311
#> [46,]   -3.048005  -0.090378755  0.02248442
#> [47,]   -2.810369  -0.075584238  0.01873685
#> [48,]   -2.093205  -0.178230683  0.01683502
#> [49,]   -3.454613   0.013795855  0.03620723
#> [50,]   -3.116492   0.068775719  0.01763898
#> 
#> $psis$s100
#>       (Intercept)          educ          age
#>  [1,] 10.24032930 -2.2688133135  0.905900143
#>  [2,]  1.23668924 -0.4577817304  0.189169307
#>  [3,]  2.04646291 -0.1837466112 -0.010533239
#>  [4,]  1.67621042 -0.0008583885 -0.042080855
#>  [5,]  0.62189508  0.0822501170 -0.034596345
#>  [6,]  0.08458480  0.1269302837 -0.027650169
#>  [7,] -1.50697937  0.3041969811 -0.014646723
#>  [8,] -1.98500929  0.3633519137 -0.022155410
#>  [9,] -2.15050329  0.2916179711 -0.012465025
#> [10,] -1.71235255  0.1926721399 -0.017941133
#> [11,] -2.60336906  0.1580859266  0.006293140
#> [12,] -2.11752724  0.0581695793  0.001305941
#> [13,] -2.04043520  0.0861976829 -0.010260214
#> [14,] -1.52875357 -0.0455236688 -0.011018042
#> [15,] -1.18135831 -0.0721741629 -0.016716285
#> [16,] -1.17373089 -0.0979851258 -0.016548651
#> [17,] -0.46667103 -0.1901438585 -0.024902937
#> [18,] -1.19726562 -0.0379825160 -0.030034280
#> [19,]  0.02820969 -0.1416997852 -0.050892071
#> [20,]  0.79128925  0.0039325418 -0.092894075
#> [21,]  1.07349875 -0.0721584424 -0.092045750
#> [22,]  2.62566542 -0.1896334036 -0.126093761
#> [23,]  1.36037652  0.0083853688 -0.114728687
#> [24,]  0.70840947  0.1350459231 -0.116116377
#> [25,] -0.23910774  0.2158799332 -0.093694689
#> [26,] -1.30699195  0.1446618996 -0.056515201
#> [27,] -0.94277957  0.1149267454 -0.063534257
#> [28,] -1.53946985  0.1063154863 -0.046094360
#> [29,] -0.13550161 -0.0181781155 -0.073067012
#> [30,] -1.35590640  0.2701932464 -0.074075767
#> [31,] -1.06710780  0.2986756712 -0.087654512
#> [32,] -0.87012736  0.2071931054 -0.084569887
#> [33,] -0.55977454  0.0969729272 -0.076292372
#> [34,] -0.48029519  0.1771468657 -0.090169013
#> [35,] -0.37043512  0.1987206816 -0.086859513
#> [36,]  1.03758501  0.0877936487 -0.115724128
#> [37,]  0.56581414  0.1647514821 -0.118135561
#> [38,]  0.80345289  0.1546164210 -0.123354537
#> [39,]  0.54744773  0.2407466095 -0.131216988
#> [40,]  0.10530164  0.1644955563 -0.104061979
#> [41,] -0.01818890  0.2351722279 -0.112720416
#> [42,] -0.56555960  0.3792883456 -0.108947813
#> [43,]  0.19067976  0.2961149962 -0.128897922
#> [44,]  1.10043473  0.2284136894 -0.147534033
#> [45,]  0.57296300  0.3265533034 -0.150391562
#> [46,] -0.01646022  0.3850728993 -0.141191549
#> [47,]  0.57190701  0.2402334306 -0.149223616
#> [48,] -0.07219741  0.3433111777 -0.139688669
#> [49,]  0.80769578  0.2127852104 -0.146543467
#> [50,]  0.96981691  0.0867615717 -0.141532260
#> 
#> $psis$s000
#>       (Intercept) educ age
#>  [1,]           0    0   0
#>  [2,]           0    0   0
#>  [3,]           0    0   0
#>  [4,]           0    0   0
#>  [5,]           0    0   0
#>  [6,]           0    0   0
#>  [7,]           0    0   0
#>  [8,]           0    0   0
#>  [9,]           0    0   0
#> [10,]           0    0   0
#> [11,]           0    0   0
#> [12,]           0    0   0
#> [13,]           0    0   0
#> [14,]           0    0   0
#> [15,]           0    0   0
#> [16,]           0    0   0
#> [17,]           0    0   0
#> [18,]           0    0   0
#> [19,]           0    0   0
#> [20,]           0    0   0
#> [21,]           0    0   0
#> [22,]           0    0   0
#> [23,]           0    0   0
#> [24,]           0    0   0
#> [25,]           0    0   0
#> [26,]           0    0   0
#> [27,]           0    0   0
#> [28,]           0    0   0
#> [29,]           0    0   0
#> [30,]           0    0   0
#> [31,]           0    0   0
#> [32,]           0    0   0
#> [33,]           0    0   0
#> [34,]           0    0   0
#> [35,]           0    0   0
#> [36,]           0    0   0
#> [37,]           0    0   0
#> [38,]           0    0   0
#> [39,]           0    0   0
#> [40,]           0    0   0
#> [41,]           0    0   0
#> [42,]           0    0   0
#> [43,]           0    0   0
#> [44,]           0    0   0
#> [45,]           0    0   0
#> [46,]           0    0   0
#> [47,]           0    0   0
#> [48,]           0    0   0
#> [49,]           0    0   0
#> [50,]           0    0   0
#> 
#> 
#> $s.tab
#>             s111       s100      s000
#>  [1,] 0.05333333 0.65333333 0.2933333
#>  [2,] 0.05333333 0.60000000 0.3466667
#>  [3,] 0.05333333 0.57333333 0.3733333
#>  [4,] 0.05333333 0.51200000 0.4346667
#>  [5,] 0.05333333 0.40266667 0.5440000
#>  [6,] 0.05333333 0.38666667 0.5600000
#>  [7,] 0.05600000 0.31733333 0.6266667
#>  [8,] 0.05600000 0.26133333 0.6826667
#>  [9,] 0.05600000 0.19466667 0.7493333
#> [10,] 0.05600000 0.17066667 0.7733333
#> [11,] 0.05600000 0.14666667 0.7973333
#> [12,] 0.05600000 0.13066667 0.8133333
#> [13,] 0.05600000 0.12266667 0.8213333
#> [14,] 0.05600000 0.10400000 0.8400000
#> [15,] 0.05600000 0.12800000 0.8160000
#> [16,] 0.05600000 0.11733333 0.8266667
#> [17,] 0.05600000 0.12266667 0.8213333
#> [18,] 0.05600000 0.09600000 0.8480000
#> [19,] 0.05600000 0.10666667 0.8373333
#> [20,] 0.05600000 0.10666667 0.8373333
#> [21,] 0.05600000 0.11733333 0.8266667
#> [22,] 0.05866667 0.12533333 0.8160000
#> [23,] 0.05866667 0.15200000 0.7893333
#> [24,] 0.05866667 0.12533333 0.8160000
#> [25,] 0.05866667 0.11200000 0.8293333
#> [26,] 0.05866667 0.11466667 0.8266667
#> [27,] 0.05866667 0.08533333 0.8560000
#> [28,] 0.05866667 0.07200000 0.8693333
#> [29,] 0.05866667 0.08800000 0.8533333
#> [30,] 0.06133333 0.08266667 0.8560000
#> [31,] 0.06666667 0.08266667 0.8506667
#> [32,] 0.05866667 0.09066667 0.8506667
#> [33,] 0.05866667 0.08000000 0.8613333
#> [34,] 0.06933333 0.07733333 0.8533333
#> [35,] 0.06933333 0.08533333 0.8453333
#> [36,] 0.07200000 0.09600000 0.8320000
#> [37,] 0.06666667 0.10133333 0.8320000
#> [38,] 0.06933333 0.11466667 0.8160000
#> [39,] 0.06933333 0.08266667 0.8480000
#> [40,] 0.06400000 0.11200000 0.8240000
#> [41,] 0.07200000 0.09600000 0.8320000
#> [42,] 0.07466667 0.08533333 0.8400000
#> [43,] 0.07200000 0.11466667 0.8133333
#> [44,] 0.06666667 0.10933333 0.8240000
#> [45,] 0.06933333 0.09600000 0.8346667
#> [46,] 0.06666667 0.08800000 0.8453333
#> [47,] 0.06933333 0.09066667 0.8400000
#> [48,] 0.07466667 0.06933333 0.8560000
#> [49,] 0.06933333 0.09066667 0.8400000
#> [50,] 0.07466667 0.08000000 0.8453333
#> 
#> $delta.1
#>  [1]  0.24225352  0.26619718  0.30140845  0.28450704  0.11830986  0.21549296
#>  [7] -0.49354318 -0.43422115 -0.50161421 -0.05891848 -0.68038741 -0.19047619
#> [13]  0.05649718 -0.15899919 -0.32728006 -0.61864407 -0.33373688 -0.35593220
#> [19] -0.53712672 -0.38337369 -0.61057304 -0.09940767  0.07391192 -0.15336080
#> [25]  0.16739634  0.34058717  0.20718517 -0.27852176 -0.34097347  0.03693182
#> [31]  0.25714286 -0.32977080  0.17048674  0.32367203  0.29953714  0.32247765
#> [37]  0.19428571  0.38461538 -0.37028874  0.11930199 -0.04214559  0.32204611
#> [43]  0.07151980  0.19142857  0.43001984  0.27428571  0.01311439  0.12958007
#> [49]  0.07857615  0.33923425
#> 
#> $delta.2
#>  [1]  0.04476431  0.21658958  0.32885535  0.22086946  0.10568854  0.15163813
#>  [7] -0.45339832 -0.40968473 -0.47973660  0.05223910 -0.72598984 -0.33607924
#> [13] -0.04233769  0.09431371 -0.35681143 -0.52546449 -0.25204547 -0.34234290
#> [19] -0.53898528 -0.39003500 -0.74212221 -0.08471972 -0.09053455 -0.20657136
#> [25]  0.44336195  0.30036799  0.30001225 -0.11913143 -0.49566291 -0.11634907
#> [31]  0.17245026 -0.43453250  0.17687101  0.33500970  0.44100699  0.23426002
#> [37]  0.21084131  0.36148367 -0.20413154  0.22443539  0.02707730  0.40214294
#> [43] -0.03085452  0.32834225  0.30612169  0.21361178  0.05798255  0.35175680
#> [49]  0.09801565  0.37053218
#>